Extract ImapFolderPusher from ImapStore
